Re-use test helper methods
