+++ programs/cli/package.json
@@ -33,7 +33,7 @@
     "compile:cli": "tsup-node ./cli.ts --format cjs --dts --target=node18",
     "compile": "yarn compile:readme-files && yarn compile:tailwind-config && yarn compile:stylelint-config &&yarn compile:cli",
     "clean": "rm -rf dist",
-    "before:test": "./spec/install-npm-deps-for-fixtures.sh",
+    "before:test": "./spec/fixtures/install-npm-deps-for-fixtures.sh",
     "test": "npm run before:test && jest"
   },
   "keywords": [
