+++ programs/cli/spec/fixtures/constants.ts
