+++ programs/cli/spec/fixtures/install-npm-deps-for-fixtures.sh
